implement scoreboard

+++ src/build.sh
@@ -3,8 +3,9 @@
 set -e
 
 print_usage() {
-	echo "Usage: $0 [-d target_directory] [--directory target_directory]"
+	echo "Usage: $0 [-d target_directory] [--directory target_directory] [-k api_key] [--api-key api_key]"
 	echo "  -d, --directory target_directory   Directory where the project will be created (default: current directory)"
+	echo "  -k, --api-key api_key              API key to communicate with the frontend application"
 }
 
 log_message() {
@@ -19,12 +20,17 @@ SCRIPT_DIR="$(dirname "$(readlink -f "$0")")"
 
 # Default values
 TARGET_DIR="."
+API_KEY=""
 
 # Parse command line options
 while [[ "$#" -gt 0 ]]; do
 	case $1 in
 	-d | --directory)
 		TARGET_DIR="$2"
+		shift 2
+		;;
+	-k | --api-key)
+		API_KEY="$2"
 		shift 2
 		;;
 	*)
@@ -35,6 +41,12 @@ while [[ "$#" -gt 0 ]]; do
 	esac
 done
 
+if [ -z "$API_KEY" ]; then
+	echo "API key is required"
+	print_usage
+	exit 1
+fi
+
 # Create target directory if it doesn't exist
 log_message "Creating target directory: $TARGET_DIR"
 rm -rf "$TARGET_DIR"
@@ -42,7 +54,7 @@ mkdir -p "$TARGET_DIR"
 
 # Build the Go program
 log_message "Building the Go program"
-GOOS=js GOARCH=wasm go build -trimpath -ldflags="-s -w" -o "$TARGET_DIR/main.wasm" "$SCRIPT_DIR/main.go"
+GOOS=js GOARCH=wasm go build -trimpath -ldflags="-s -w -X 'main.ApiKey=$API_KEY'" -o "$TARGET_DIR/main.wasm" "$SCRIPT_DIR/main.go"
 
 if [ -f "$TARGET_DIR/main.wasm" ]; then
 	log_message "Go program built successfully"
@@ -52,7 +64,7 @@ else
 fi
 
 # Download the Go runtime for WebAssembly
-log_message "Downloading the Go runtime for WebAssembly"
+log_message "Retrieving the Go runtime for WebAssembly"
 if [ -f "$(go env GOROOT)/misc/wasm/wasm_exec.js" ]; then
 	cp "$(go env GOROOT)/misc/wasm/wasm_exec.js" "$TARGET_DIR/wasm_exec.js"
 	log_message "Go runtime copied successfully"
